backend/requirements.txt
Django==5.0.1
django-environ==0.11.2
gunicorn==21.2.0
psycopg==3.1.17
psycopg-binary==3.1.17
